% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Spatial_pre_processing.R
\name{add_density_attributes_to_data}
\alias{add_density_attributes_to_data}
\title{Add density attributes to data}
\usage{
add_density_attributes_to_data(crime_data, crime_kde, population_kde)
}
\arguments{
\item{crime_data}{full crime dataset}

\item{crime_kde}{kernel density estimate of crimes}

\item{population_kde}{kernel density estimate of population}
}
\value{
crime dataset with three additional columns:
\item{\code{Crime Density}}{crime density at the location of crime}
\item{\code{Population Density}}{population density at the location of crime}
\item{\code{Ratio Density}}{crime/population density at the location of crime}
}
\description{
Add density attributes to data
}
