% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Spatial_pre_processing.R
\name{add_location_to_population_data}
\alias{add_location_to_population_data}
\title{Add Location Data}
\usage{
add_location_to_population_data(block_boundaries_data, population_data)
}
\arguments{
\item{block_boundaries_data}{dataset of census block boundaries}

\item{population_data}{dataset of population by census block}
}
\value{
population dataset with location data
}
\description{
Adds locations information to population data from census block id
}
