% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{all_crime}
\alias{all_crime}
\alias{sub_all_crime}
\title{Chicago Crime data from 2001 to 2019}
\format{A data frame with 6,336,525 rows and 14 variables:
\describe{
  \item{Date}{date of which the crime was reported}
  \item{Primary Type}{type of crime that was reported, one of a finite amount of options}
  \item{Description}{description of the reported crime}
  \item{Location Description}{type of place where the crime was committed}
  \item{Arrest}{logical; if \code{TRUE}, then the crime resulted in an arrest}
  \item{Domestic}{logical; if \code{TRUE}, then the incident was domestic-related}
  \item{Beat}{the beat where the crime occurred (smallest police geographical area)}
  \item{District}{police district where the crime occurred}
  \item{Ward}{the ward (City Council district) in Chicago where the crime occurred}
  \item{Community Area}{the community area where the incident occurred}
  \item{Year}{year when the crime happened}
  \item{Latitude}{latitude co-ordinate (in degrees) of the block where the crime occurred}
  \item{Longitude}{longitude co-ordinate (in degrees) of the block where the crime occurred}
  \item{Hour}{estimated hour of the day when the crime happened}
}}
\source{
\url{https://data.cityofchicago.org/Public-Safety/Crimes-2001-to-present/ijzp-q8t2}
}
\usage{
data(all_crime)

data(sub_all_crime)
}
\description{
A dataset containing reported crimes in the city of Chicago in the period 2001 to 2019.
}
\keyword{datasets}
