% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LR_OOP.R
\name{auc_in}
\alias{auc_in}
\title{AUC Calculator}
\usage{
auc_in(p, y)
}
\arguments{
\item{p}{a vector of predictions}

\item{y}{a vector of the true observed response variable}
}
\value{
the AUC value
}
\description{
See \code{\link{roc.lr}} for details. This function is used to only output the AUC value for use in other functions, such as \code{\link{cv.lr}}.
}
