% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{block_boundaries}
\alias{block_boundaries}
\title{Chicago Census Block Boundaries}
\format{A data frame with 46291 rows and 3 variables:
\describe{
  \item{the_geom}{string indicating polygon co-ordinates of the census block}
  \item{STATEFP10}{two digit number referring to state code of census block}
  \item{COUNTYFP10}{three digit number refferring to county code of census block}
  \item{TRACTCE10}{six digit number referring to area in chicago of census block}
  \item{BLOCKCE10}{four digit number referring to block in census block}
  \item{GEOID10}{full identification number of census block, same as \code{CENSUS BLOCK FULL} in \code{\link{block_boundaries}}}
  \item{NAME10}{name of census block}
  \item{TRACT_BLOC}{shortened identification number of the census block, without the first 5 numbers (relating to area in country), same as \code{CENSUS BLOCK} in \code{\link{block_boundaries}}}

}}
\source{
\url{https://data.cityofchicago.org/Facilities-Geographic-Boundaries/Boundaries-Census-Blocks-2010/mfzt-js4n}
}
\usage{
block_boundaries
}
\description{
A dataset containing the locations of census blocks in Chicago
}
\keyword{datasets}
