% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{block_populations}
\alias{block_populations}
\title{Chicago Population data by Census block}
\format{A data frame with 46291 rows and 3 variables:
\describe{
  \item{CENSUS BLOCK}{shortened identification number of the census block, without the first 5 numbers (relating to area in country), same as \code{TRACT_BLOC} in \code{\link{block_populations}}}
  \item{CENSUS BLOCK FULL}{full identification number of the census block, same code as \code{GEOID10} in \code{\link{block_populations}}}
  \item{TOTAL POPULATION}{total population per census block}
}}
\source{
\url{https://data.cityofchicago.org/Facilities-Geographic-Boundaries/Population-by-2010-Census-Block/5yjb-v3mj}
}
\usage{
block_populations
}
\description{
A dataset containing the population totals of census blocks in the city of Chicago
}
\keyword{datasets}
