% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{census}
\alias{census}
\title{Chicago Census data from 2008-2012}
\format{A data frame with 78 rows and 9 variables:
\describe{
  \item{Community Area Number}{identification number of the community area, matches with that in \code{all_crime}}
  \item{COMMUNITY AREA NAME}{name of the community area}
  \item{PERCENT OF HOUSING CROWDED}{Percentage of occupied housing units, in the community area, with more than one person per room}
  \item{PERCENT HOUSEHOLDS BELOW POVERTY}{Percentage of households, in the community area, that are living below the federal poverty level }
  \item{PERCENT AGED 16+ UNEMPLOYED}{Percentage of persons, in the community area, that are over the age of 16 years that are unemployed}
  \item{PERCENT AGED 25+ WITHOUT HIGH SCHOOL DIPLOMA}{Percentage of persons in the community area that are over the age of 25 years without a high school education}
  \item{PERCENT AGED UNDER 18 OR OVER 64}{Percent of the population of the community area under 18 or over 64 years of age}
  \item{PER CAPITA INCOME}{Community Area Per capita income; estimated as the sum of tract-level aggregate incomes divided by the total population }
  \item{HARDSHIP INDEX}{Score that incorporates each of the six selected socioeconomic indicators}
}}
\source{
\url{https://data.cityofchicago.org/Health-Human-Services/Census-Data-Selected-socioeconomic-indicators-in-C/kn9c-c2s2}
}
\usage{
census
}
\description{
A dataset containing selected socioeconomic indicators in the city of Chicago from 2008 to 2012
}
\keyword{datasets}
