% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Spatial_pre_processing.R
\name{create_spatial_attributes}
\alias{create_spatial_attributes}
\title{Create spatial attributes}
\usage{
create_spatial_attributes(crime_data, block_boundaries_data, population_data,
  h = 0.01, grid.size = 100L)
}
\arguments{
\item{crime_data}{full crime dataset}

\item{block_boundaries_data}{dataset of census block boundaries}

\item{population_data}{dataset of population by census block}

\item{h}{bandwidth for kernel density estimation}

\item{grid.size}{grid size for kernel density estimation}
}
\value{
dataset with attributes added
}
\description{
Create spatial attributes
}
