% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Spatial_pre_processing.R
\name{fit_kde}
\alias{fit_kde}
\title{Fit Kernel Density Estimator}
\usage{
fit_kde(data, h, grid.size)
}
\arguments{
\item{data}{Data frame containing columns \code{Longitude} and \code{Latitude}.}

\item{h}{Bandwidth of KDE.}

\item{grid.size}{Number of rows/columns of grid on which density estimate is returned.}
}
\value{
Kernel density estimate in the form:
\item{\code{x1}}{Longitudes of grid}
\item{\code{x2}}{Latitudes of grid}
\item{\code{fhat}}{Matrix of density estimates at grid points}
}
\description{
Fits Kernel density estimates using \code{\link[KernSmooth]{bkde2D}}
}
