% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_data.R
\name{formatData}
\alias{formatData}
\title{Format crime data to include census and police station data}
\usage{
formatData(crime_data, densities = TRUE, verbose = TRUE, ...)
}
\arguments{
\item{crime_data}{dataframe from \code{data(all_crime)}}

\item{densities}{logical; if \code{TRUE}, population densities, crime densities, as well as the ratio between the two will also be joined to \code{crime_data}}

\item{verbose}{logical; if \code{TRUE}, formatting procedures will be displayed as they are completed}

\item{...}{further arguments to be passed to \code{\link{create_spatial_attributes}}}
}
\value{
data of the same form of \code{crime_data}, with appended census and police station data
}
\description{
Format crime data to include census and police station data
}
\examples{
\dontrun{
data(sub_all_crime)
sub_all_crime = cleanData(sub_all_crime)
sub_all_crime = formatData(sub_all_crime) }
}
