% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LR_OOP.R
\name{get_error}
\alias{get_error}
\title{Match Error Function}
\usage{
get_error(type = "mse")
}
\arguments{
\item{type}{type of error function to be output, either "mse", "auc" or "log"}
}
\value{
function of either mean squared error or \code{\link{auc_in}}
}
\description{
Used to get error function in \code{\link{cv.lr}}.
}
