% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LR_OOP.R
\name{irls.lr}
\alias{irls.lr}
\title{Iteratively Re-weighted Least Squares}
\usage{
irls.lr(y, x, init = NULL, tol = 1e-06, maxiter = 100)
}
\arguments{
\item{y}{response vector}

\item{x}{model matrix of covariates}

\item{init}{initial estimate of \code{theta}}

\item{tol}{tolerance parameter, default \code{1e-6}}

\item{maxiter}{optional number of iterations}
}
\value{
a list containing
\item{\code{par}}{a vector of estimates of \code{theta}, the parameters being optimised}
\item{\code{val}}{the value of the log-likelihood at the final \code{theta} estimate}
\item{\code{iters}}{the number of iterations needed to converge}
}
\description{
Optimisation procedure based on iteratively re-weighted least squares, called by \code{\link{lr}}.
}
\details{
Iterative method used to find parameter estimates of \eqn{\beta} from the least squares problem
\deqn{
\beta = argmin (z - X\beta)^T W (z- X\beta),
}
where \eqn{W} is a diagonal matrix of weights with \eqn{i}-th diagonal element being
\deqn{
\sigma(x_i ; \beta) (1 - \sigma(x_i ; \beta)),
}
and \eqn{z} is the vector
\deqn{
z = X \beta + W^{-1} (y - \sigma(x_i ; \beta)).
}
The parameter vector \eqn{\beta} is updated iteratively with a Newton update of the form
\deqn{
\beta = (X^T W X)^{-1} X^T W z.
}
}
