% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LR_OOP.R
\name{loglik_lr}
\alias{loglik_lr}
\title{Bernoulli distribution log-likelihood}
\usage{
loglik_lr(theta, x, y)
}
\arguments{
\item{theta}{parameters relating to \code{x}}

\item{x}{explanatory variables in model matrix}

\item{y}{binary response variable}
}
\value{
single value, the log-likelihood of the Bernoulli distribution with fixed \code{x} and \code{y}
}
\description{
Log-likelihood of the Bernoulli distribution, commonly used in optimisation procedures for maximisation.
}
\details{
The log-likelihood is written
\deqn{
\sum_{i=1}^n y \log(\sigma(X\beta)) + (1-y) \log(1-\sigma(X\beta))
}
}
\examples{
theta = c(1, 0.1)
X = matrix(rnorm(4), 50, 2)
y = sample(0:1, 50, replace=TRUE)
loglik_lr(theta, X, y)
}
