% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LR_OOP.R
\name{lr}
\alias{lr}
\title{Logistic Regression}
\usage{
lr(formula, data, init = NULL)
}
\arguments{
\item{formula}{an object of class \code{\link[stats]{formula}}, a symbolic description of the model to be fitted.
The specified names need to also be in \code{data}.}

\item{data}{a required data frame containing the variables in the model}

\item{init}{optional initial conditions to be passed to optimisation of the log-likelihood}
}
\value{
An S3 object of class 'lr', which is a list containing
\item{\code{coefficients}}{a vector of coefficients corresponding to covariates specified in \code{formula}}
\item{\code{data}}{the \code{data} input to the function}
\item{\code{formula}}{the \code{formula} input to the function}
\item{\code{X}}{the model matrix \code{X}}
\item{\code{val}}{value of the final log-likelihood at the values of \code{coefficients}, given by \code{loglik_lr}}
\item{\code{its}}{number of iterations performed to retrieve the maximised log-likelihood}
}
\description{
\code{lr} is used to fit a logistic regression model for a binary response variable.
}
\details{
The form of the \code{formula} argument will be of the form \code{response ~ predictor1 + predictor2 + ...},
with \code{predictor1} and \code{predictor2} being named columns of the data frame in \code{data}.

The log-likelihood (from \code{loglik_lr}) is maximised using \code{\link{irls.lr}} with initial estimates given by \code{init}.
If no initial values are supplied, this uses a vector of zeros instead.
}
\examples{
y = sample(0:1, 50, replace=TRUE)
d = data.frame(y = y, x = rnorm(10*y + 15))
fit = lr(y ~ x, data = d)
}
