% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Spatial_pre_processing.R
\name{make_population_data_for_kde}
\alias{make_population_data_for_kde}
\title{Format population data}
\usage{
make_population_data_for_kde(population_data)
}
\arguments{
\item{population_data}{dataset of populations at locations}
}
\value{
dataset of home locations with one row per person
}
\description{
Make population data suitable for \code{\link{fit_kde}} function
}
