% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Spatial_pre_processing.R
\name{plot_kde}
\alias{plot_kde}
\title{Plot Kernel Density Estimate}
\usage{
plot_kde(kde)
}
\arguments{
\item{kde}{Kernel density estimate object generated using \code{\link{fit_kde}}}
}
\description{
Plots kernel density estimates on map using \code{\link[leaflet]{leaflet}}
}
