% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{police_stations}
\alias{police_stations}
\title{Chicago Police Station locations}
\format{A data frame with 23 rows and 2 variables:
\describe{
  \item{lon}{longitude of a police station}
  \item{lat}{latitude of a police station}
}}
\source{
\url{https://data.cityofchicago.org/Public-Safety/Police-Stations/z8bn-74gv}
}
\usage{
police_stations
}
\description{
A dataset containing the longitudes and latitudes of police stations in the city of Chicago, updated June 2016
}
\keyword{datasets}
