% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LR_OOP.R
\name{predict.lr}
\alias{predict.lr}
\title{Predict from a Logistic Regression model}
\usage{
\method{predict}{lr}(object, newdata = NULL, thresh = 0, type = "preds",
  ...)
}
\arguments{
\item{object}{an object of class "\code{lr}", the output from \code{\link{lr}}}

\item{newdata}{an optional data frame to predict from. If omitted, predictions will be from the original data frame used to fit the model.}

\item{thresh}{an optional threshold parameter, see 'Details'.}

\item{type}{type of predictions, can be one of "preds", "probs" or "vals", see 'Details'.}

\item{...}{further arguments}
}
\value{
a vector of predictions, probabilities or values, depending on the input to \code{type}.
}
\description{
Predicted values of a certain type based on a \code{\link{lr}} object.
}
\details{
Predictions from a logistic regression model are formed by first multiplying the model matrix by the parameter vector, i.e.
\deqn{
f(x; \beta) := X \beta.
}
These values are outputted when the \code{type} argument is "vals". Predictions for the positive class can be obtained when these values
are above a certain threshold value, the argument \code{thresh}, i.e.
\deqn{
f(x; \beta) \ge t,
}
where the threshold \eqn{t} is usually equal to zero. This is how the predictions are made when the \code{type} argument is "preds".
To get probabilities, the \code{\link{sigmoid}} function is used on the values, i.e.
\deqn{
p(y=1) = 1/(1 + e^{-X \beta}).
}
}
\examples{
y = sample(0:1, 50, replace=TRUE)
d = data.frame(y = y, x = rnorm(10*y + 15))
fit = lr(y ~ x, data = d)
predict(fit, type="preds")
}
