% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LR_OOP.R
\name{print.lr}
\alias{print.lr}
\title{Print Logistic Regression Model}
\usage{
\method{print}{lr}(x, ...)
}
\arguments{
\item{x}{an object of class "\code{lr}", the output from \code{\link{lr}}}

\item{...}{further arguments}
}
\value{
a printed, named vector of coefficients and log-likelihood value at these estimates
}
\description{
Displays important output from a \code{\link{lr}} object; the parameter estimates and the maximised log-likelihood value.
}
\examples{
y = sample(0:1, 50, replace=TRUE)
d = data.frame(y = y, x = rnorm(10*y + 15))
fit = lr(y ~ x, data = d)
print(fit)
}
