% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Spatial_pre_processing.R
\name{process_geom_string}
\alias{process_geom_string}
\title{Geom string process
Process \code{the_geom} string to return block center}
\usage{
process_geom_string(x, long = TRUE)
}
\arguments{
\item{x}{\code{the_geom} string from block boundaries data}

\item{long}{if TRUE returns longitude, else returns latitide}
}
\value{
center of block (longitude or latitude)
}
\description{
Geom string process
Process \code{the_geom} string to return block center
}
