% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LR_OOP.R
\name{roc.lr}
\alias{roc.lr}
\title{Receiver Operating Characteristic (ROC) curve}
\usage{
roc.lr(lrfit, newdata = NULL, plot = TRUE, len = 50)
}
\arguments{
\item{lrfit}{an object of class "\code{lr}", the output from \code{\link{lr}}}

\item{newdata}{an optional data frame to predict from. If ignored, the default data frame is that used to fit the original model.}

\item{plot}{logical; if \code{TRUE}, then the ROC curve is plotted}

\item{len}{optional; number of different thresholds to use.}
}
\value{
the AUC value, and a plot of the ROC curve if \code{plot=TRUE}
}
\description{
A method of judging the predictive performance of a model by plotting and/or averaging the
probability of predicting the positive class correctly, over multiple thresholds. See 'Details'.
}
\details{
A positive prediction from a logistic regression model is made when
\deqn{
f(x; \beta) := X \beta \ge t.
}

where \eqn{t} is some threshold. See \code{\link{predict.lr}} for details. A different threshold
\eqn{t_0} will yield a different set of predictions. For a given sequence \eqn{t_j in [min(t), max(t)]},
 for \eqn{j=1,\dots,J}, the True Positive Rate (TPR) and False Positive Rate (FPR) can be calculated as
\deqn{
 TPR(j) = \sum I(f(x_i;\beta) \geq t_j)/\sum I(y_i = 1),
}
\deqn{
 FPR(j) = \frac{\sum I(f(x_i;\beta) < t_j)}{\sum I(y_i = 1)}.
}
The ROC curve is plotted from the pairs \eqn{(FPR(j), TPR(j))}, and the AUC is calculated as the area under this curve, i.e.
\deqn{
 AUC = \int_{j=1}^J TPR(FPR(j))dj.
}
}
