% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LR_OOP.R
\name{sigmoid}
\alias{sigmoid}
\title{Sigmoid function}
\usage{
sigmoid(z)
}
\arguments{
\item{z}{input}
}
\value{
output value of the function
}
\description{
Sigmoid function, otherwise known as a logistic function
}
\details{
The sigmoid (or logistic) function,
\deqn{
\sigma(z) = 1 / (1 + e^{-z})
}
is used in logistic regression to model probabilities, commonly the probability of predicting the positive class,
i.e. \eqn{p(y=1)}.
}
\examples{
theta = c(1, 0.1)
X = matrix(rnorm(4), 2, 2)
sigmoid(X \%*\% theta)
}
